+++ lib/roulette.dart
@@ -16,8 +16,13 @@
 /// @author do9core
 library roulette;
 
-export './src/roulette.dart' show Roulette;
-export './src/roulette_controller.dart' show RouletteController;
+export './src/roulette.dart' show Roulette, TappableRoulette;
+export './src/roulette_controller.dart'
+    show
+        RouletteController,
+        AnimationConfig,
+        CurveAnimationConfig,
+        PhysicsAnimationConfig;
 export './src/roulette_group.dart' show RouletteGroup;
 export './src/roulette_unit.dart' show RouletteUnit;
-export './src/roulette_style.dart' show RouletteStyle;
+export './src/roulette_style.dart' show RouletteStyle, SectionImageLayout;